requirements.txt
Flask==2.2.2
Werkzeug==2.2.2  # Specifically specify a compatible version
pymongo==4.3.3
python-dotenv==0.21.1
black==22.12.0
